+++ pkg/upload/android-proguard.go
@@ -28,7 +28,8 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 	proguardOptions := options.Upload.AndroidProguard
 
 	var mappingFile string
-	var appManifestPathExpected string
+	var primaryAppManifestPath string
+	var fallbackAppManifestPath string
 	var err error
 
 	for _, path := range proguardOptions.Path {
@@ -57,10 +58,14 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 
 			// Attempt to locate AndroidManifest.xml for the variant if not set
 			if proguardOptions.AppManifest == "" {
-				appManifestPathExpected = filepath.Join(path, "app", "build", "intermediates", "merged_manifests", proguardOptions.Variant, "AndroidManifest.xml")
-				if utils.FileExists(appManifestPathExpected) {
-					proguardOptions.AppManifest = appManifestPathExpected
-					logger.Info(fmt.Sprintf("Found app manifest at: %s", proguardOptions.AppManifest))
+				primaryAppManifestPath = filepath.Join(path, "app", "build", "intermediates", "merged_manifests", proguardOptions.Variant, "AndroidManifest.xml")
+
+				fallbackAppManifestPath = filepath.Join(path, "app", "build", "intermediates", "merged_manifests", proguardOptions.Variant, "process"+proguardOptions.Variant+"Manifest", "AndroidManifest.xml")
+
+				proguardOptions.AppManifest = android.FindAndroidManifest([]string{primaryAppManifestPath, fallbackAppManifestPath})
+
+				if proguardOptions.AppManifest == "" {
+					logger.Info("Unable to locate AndroidManifest.xml for variant; proceeding without manifest path")
 				}
 			}
 
@@ -74,10 +79,14 @@ func ProcessAndroidProguard(options options.CLI, logger log.Logger) error {
 				if filepath.Base(mergedManifestPath) == "merged_manifests" {
 					proguardOptions.Variant, err = android.GetVariantDirectory(mergedManifestPath)
 					if err == nil {
-						appManifestPathExpected = filepath.Join(mergedManifestPath, proguardOptions.Variant, "AndroidManifest.xml")
-						if utils.FileExists(appManifestPathExpected) {
-							proguardOptions.AppManifest = appManifestPathExpected
-							logger.Info(fmt.Sprintf("Found app manifest at: %s", proguardOptions.AppManifest))
+						primaryAppManifestPath = filepath.Join(mergedManifestPath, proguardOptions.Variant, "AndroidManifest.xml")
+
+						fallbackAppManifestPath = filepath.Join(mergedManifestPath, proguardOptions.Variant, "process"+proguardOptions.Variant+"Manifest", "AndroidManifest.xml")
+
+						proguardOptions.AppManifest = android.FindAndroidManifest([]string{primaryAppManifestPath, fallbackAppManifestPath})
+
+						if proguardOptions.AppManifest == "" {
+							logger.Info("Unable to locate AndroidManifest.xml for variant; proceeding without manifest path")
 						}
 					}
 				}